requirements.txt
python-dotenv
discord.py
transformers
torch